[Enhancement] Rework of the imbuement system and conversion to c++ (#339)
NOTE: The pull request of optimization and enhancement of ItemAttributes is needed for this pull request: #337. If don't use this pull, the imbuement will crash!
Thanks to @lgrossi for the reviews made and for helping to solve several problems.
Thanks @Glatharth for provide the script for convert tables lua in items.xml: https://gist.github.com/dudantas/548fdbc3c521327c7d5320e27adba8fb

The system has been completely converted to c++
Fixed bugs and adapted the operation to be the same as global tibia
Possibility to imbue the equipped item
Added config.lua key for enable/disable the use of storages by the imbuement system (togleImbuementShrineStorage)

The imbuement of each item is now configured in items.xml in the following settings:
<attribute key="imbuementslot" value="1">
	<attribute key="elemental damage" value ="1"/>
	<attribute key="life leech" value ="1"/>
	<attribute key="mana leech" value ="1"/>
	<attribute key="critical hit" value ="1"/>
	<attribute key="elemental protection death" value ="1"/>
	<attribute key="elemental protection earth" value ="1"/>
	<attribute key="elemental protection fire" value ="1"/>
	<attribute key="elemental protection ice" value ="1"/>
	<attribute key="elemental protection energy" value ="1"/>
	<attribute key="elemental protection holy" value ="1"/>
	<attribute key="increase speed" value ="1"/>
	<attribute key="skillboost axe" value ="1"/>
	<attribute key="skillboost sword" value ="1"/>
	<attribute key="skillboost club" value ="1"/>
	<attribute key="skillboost shielding" value ="1"/>
	<attribute key="skillboost distance" value ="1"/>
	<attribute key="skillboost magic level" value ="1"/>
	<attribute key="increase capacity" value ="1"/>
</attribute>

Where the value of the imbuementslot goes from 1 to 3 and the values of the imbuement names are always 1, you should put the name as above, as it is from the name that the value of each imbuement is called by the "loadFromXML" function.