remove addEvent
Concerns: I don't know why this store uses addEvent for everything.
Unfortunately I couldn't fix the addEvent because when I sent the uniqueId to the addEvent, later I cannot retrieve the item through Item(uid) which is VERY STRANGE.

So I'm removing the addEvent but be aware that this might introduce a situation where you can buy the item with packet editor and then closes the store and log out before the coins are removed from your account. (Potentially) or even get the item created before the store checks you're not with it open.