Fixed questlog tracker to best follow missions with counters (#605)
With quests using counters like Killing In the Name of, the number of kills was not being updated dynamically on every kill, in order to update the tracker you must disable/enable the tracker for any mission. With this change you can simple add the same value to the questline and it will push the tracker to update and bring the new counters values.